API Reference

Index

SingleDealerAssortment.Coaml_model

SingleDealerAssortment.Coaml_model

SingleDealerAssortment.Dataset

SingleDealerAssortment.Dataset

SingleDealerAssortment.Instance

SingleDealerAssortment.Results

SingleDealerAssortment.Scenario

SingleDealerAssortment.Solution

SingleDealerAssortment.TrainingResults

SingleDealerAssortment.XSample

SingleDealerAssortment.add_policy_results!

SingleDealerAssortment.add_solution_to_dataset!

SingleDealerAssortment.build_Coaml_model

SingleDealerAssortment.build_Coaml_model_relu

SingleDealerAssortment.build_Coaml_model_tanh

SingleDealerAssortment.check_solution

SingleDealerAssortment.compute_bigM

SingleDealerAssortment.compute_cost

SingleDealerAssortment.compute_dol_archetype

SingleDealerAssortment.compute_gap_policy

SingleDealerAssortment.compute_physical_stock

SingleDealerAssortment.compute_step_instance

SingleDealerAssortment.compute_step_solution

SingleDealerAssortment.compute_μ_σ

SingleDealerAssortment.convert_milp_solution

SingleDealerAssortment.create_archetype_features

SingleDealerAssortment.create_stock_features

SingleDealerAssortment.create_x_sample

SingleDealerAssortment.g

SingleDealerAssortment.g_without_η

SingleDealerAssortment.generate_dataset

SingleDealerAssortment.generate_random_instance

SingleDealerAssortment.get_global_policy_metric

SingleDealerAssortment.get_mean_std_per_archetype_policy_metric

SingleDealerAssortment.get_policy_results

SingleDealerAssortment.get_scenarios

SingleDealerAssortment.h

SingleDealerAssortment.highs_model

SingleDealerAssortment.mean_dols_per_archetype

SingleDealerAssortment.mean_gap_model

SingleDealerAssortment.mean_loss

SingleDealerAssortment.mean_replenishment_per_archetype

SingleDealerAssortment.mean_sales_per_archetype

SingleDealerAssortment.mean_stock_per_archetype

SingleDealerAssortment.merge_dataset

SingleDealerAssortment.normalize_data!

SingleDealerAssortment.order_policies!

SingleDealerAssortment.print_costs

SingleDealerAssortment.print_gaps

SingleDealerAssortment.reduce_data!

SingleDealerAssortment.replenishment_problem

SingleDealerAssortment.sample_scenario

SingleDealerAssortment.solve_anticipative

SingleDealerAssortment.train_model

SingleDealerAssortment.train_setup

SingleDealerAssortment.transform_data_for_training_set

Reference

SingleDealerAssortment.Coaml_model — Type

struct Coaml_model{L1, L2}

Fields

θ_model::Any: replenishment reward

η_model::Any: stock penalization

SingleDealerAssortment.Coaml_model — Method

Coaml_model(
    θ_model,
    η_model
) -> SingleDealerAssortment.Coaml_model

SingleDealerAssortment.Dataset — Type

mutable struct Dataset

Fields

instance::Instance

X::Vector{SingleDealerAssortment.XSample}

Y::Vector{Any}

Y_gap::Vector{Solution}

SingleDealerAssortment.Dataset — Method

Dataset(
    instance::Instance,
    X::Vector{SingleDealerAssortment.XSample},
    Y::Vector{Any},
    Y_gap::Vector{Solution}
) -> Dataset

Create a dataset given its fields.

SingleDealerAssortment.Instance — Type

mutable struct Instance

Fields

T::Int64

n::Int64

unormalized_features::Matrix{Float64}

archetype_features::Matrix{Float64}

constraints_matrix::Vector{Vector{Float64}}

quotas::Vector{Vector{Float64}}

stock_ini::Vector{Int64}

stock_inf::Int64

stock_sup::Int64

over_stock_bound_cost::Float16

prices::Vector{Float16}

virtual_stock_costs::Vector{Float16}

physical_stock_costs::Vector{Float16}

ub_same_archetype::Int64

date_mada::Int64

min_quota_per_time_step_per_archetype::Vector{Vector{Float64}}

nb_constraints::Int64

nb_features::Int64

SingleDealerAssortment.Results — Type

mutable struct Results

Fields

instance::Instance

scenarios::Vector{Scenario}

policies::Vector{String}

solutions::Dict{String, Vector{Solution}}

SingleDealerAssortment.Scenario — Type

mutable struct Scenario

Fields

T::Int64

nb_customer::Vector{Int64}

static_utilities::Vector{Float64}

utilities::Vector{Vector{Vector{Float64}}}

customer_choice_model::Distributions.Distribution

sorted_utilities::Vector{Vector{Vector{Int64}}}

index_no_buy::Vector{Vector{Int64}}

big_M::Vector{Vector{Vector{Int64}}}

SingleDealerAssortment.Solution — Type

mutable struct Solution

Fields

instance::Instance

scenario::Scenario

stock::Vector{Vector{Int64}}

replenishments::Vector{Vector{Int64}}

sales::Vector{Vector{Int64}}

cost::Union{Nothing, Float64}

physical_stock::Union{Nothing, Vector{Vector{Int64}}}

elapsed_time::Union{Nothing, Float64}

SingleDealerAssortment.TrainingResults — Type

mutable struct TrainingResults

Fields

initial_model::Any

instance::Instance

training_set::Dataset

test_set::Dataset

final_model::Any

best_model::Any

training_losses::Vector{Float64}

validation_losses::Vector{Float64}

training_gaps::Vector{Float64}

validation_gaps::Vector{Float64}

best_validation_loss::Float64

nb_epochs::Int64

SingleDealerAssortment.XSample — Type

mutable struct XSample

Fields

instance::Instance

current_solution::Solution

features_archetypes::Matrix{Float64}

features_stock::Matrix{Float64}

SingleDealerAssortment.add_policy_results! — Method

add_policy_results!(
    results::Results,
    policy::String,
    solutions_policy::Vector{Solution}
) -> Vector{Solution}

Add results for a given policy.

SingleDealerAssortment.add_solution_to_dataset! — Method

add_solution_to_dataset!(
    dataset::Dataset,
    solution::Solution
)

Add a solution to the dataset by extracting the corresponding samples.

SingleDealerAssortment.build_Coaml_model — Method

build_Coaml_model(
    archeytpe_feature_size,
    stock_archetype_feature_size
) -> SingleDealerAssortment.Coaml_model

SingleDealerAssortment.build_Coaml_model_relu — Method

build_Coaml_model_relu(
    archeytpe_feature_size,
    stock_archetype_feature_size
) -> SingleDealerAssortment.Coaml_model

SingleDealerAssortment.build_Coaml_model_tanh — Method

build_Coaml_model_tanh(
    archeytpe_feature_size,
    stock_archetype_feature_size
) -> SingleDealerAssortment.Coaml_model

SingleDealerAssortment.check_solution — Method

check_solution(solution) -> Bool

Check if a given solution is feasible.

SingleDealerAssortment.compute_bigM — Method

compute_bigM(
    scenario::Scenario,
    instance::Instance
) -> Vector{Vector{Vector{Int64}}}

Compute big M values for the scenario and instance.

SingleDealerAssortment.compute_cost — Method

compute_cost(
    solution::Solution
) -> Union{Nothing, Float16, Float64}

Compute the cost of a given solution.

SingleDealerAssortment.compute_dol_archetype — Method

compute_dol_archetype(
    solution::Solution,
    archetype_index::Int64
) -> Union{Vector{Any}, Vector{Float64}, Vector{Int64}}

Compute the days on lot of the archetype archetype_index. It represents the time spent at the dealership by this archetype before being sold.

SingleDealerAssortment.compute_gap_policy — Method

compute_gap_policy(
    results::Results,
    policy::String
) -> Vector{Any}

Compute gap to the anticipative bound for each policy.

SingleDealerAssortment.compute_physical_stock — Method

compute_physical_stock(
    solution::Solution
) -> Union{Nothing, Vector{Vector{Int64}}}

Compute physical stock from solution.

SingleDealerAssortment.compute_step_instance — Method

compute_step_instance(
    instance::Instance,
    stock_ini::Vector{Int64},
    t::Int64
) -> Instance

Compute the step instance at time t given the initial stock.

SingleDealerAssortment.compute_step_solution — Method

compute_step_solution(
    solution::Solution,
    t::Int64
) -> Solution

Return the state of the solution at time step t.

SingleDealerAssortment.compute_μ_σ — Method

compute_μ_σ(X::Vector{Instance})

Compute mean and standard deviation for each features in X instances.

SingleDealerAssortment.convert_milp_solution — Method

convert_milp_solution(
    instance,
    scenario,
    s_val,
    y_val,
    α_val
) -> Solution

Convert MILP solution variables into Solution struct.

SingleDealerAssortment.create_archetype_features — Method

create_archetype_features(
    instance::Instance,
    step_solution::Solution
) -> Matrix{Float64}

Create features per archetype. The first instance.nb_features columns correspond to static features. The last 6 columns correspond to dynamic features:

current stock and scaled with price

mean sales and scaled with price

mean days on lot and scaled with price

SingleDealerAssortment.create_stock_features — Method

create_stock_features(
    instance::Instance,
    step_solution::Solution,
    features_archetypes
) -> Any

Create features per stock level per archetype. The first instance.nb_features+6 columns correspond to static the archetype features. The last 8 columns correspond to dynamic stock features:

deviation from stock_inf and scaled with price

deviation from stock_sup and scaled with price

deviation from min_quota and i and scaled with price

deviation from mean stock and scaled with price

SingleDealerAssortment.create_x_sample — Method

create_x_sample(
    instance::Instance,
    step_solution::Solution,
    t::Int64
) -> SingleDealerAssortment.XSample

Create input for the coaml model givent the step solution at time t.

SingleDealerAssortment.g — Method

g(y; x)

SingleDealerAssortment.g_without_η — Method

g_without_η(y; x)

SingleDealerAssortment.generate_dataset — Method

generate_dataset(
    nb_scenarios::Int64,
    instance::Instance,
    static_utilities::Vector{Float64},
    customer_choice_model::Distributions.Distribution,
    seed::Int64;
    training
) -> Dataset

Generate a dataset of scenarios with their corresponding optimal replenishment decisions.

SingleDealerAssortment.generate_random_instance — Method

generate_random_instance(
    T::Int64,
    n::Int64;
    nb_constraints,
    nb_features_archetypes,
    stock_inf,
    stock_sup,
    date_mada,
    ub_same_archetype,
    seed
) -> Instance

Generate a random instance of the problem.

SingleDealerAssortment.get_global_policy_metric — Method

get_global_policy_metric(
    results::Results,
    policy::String,
    metric::String
) -> AbstractVector{<:Union{Nothing, AbstractFloat}}

Compute global metric for a given policy and metric.

SingleDealerAssortment.get_mean_std_per_archetype_policy_metric — Method

get_mean_std_per_archetype_policy_metric(
    results::Results,
    policy::String,
    metric::String
) -> Tuple{Vector, Vector}

Compute per archetype metric for a given policy and metric.

SingleDealerAssortment.get_policy_results — Method

get_policy_results(
    results::Results,
    policy::String
) -> Vector{Solution}

Get the list of solutions for a given policy.

SingleDealerAssortment.get_scenarios — Method

get_scenarios(dataset::Dataset) -> Vector{Scenario}

Get the scenarios corresponding to the solutions stored in the dataset.

SingleDealerAssortment.h — Method

h(y; x)

SingleDealerAssortment.highs_model — Method

Initialize a HiGHS model (with disabled logging).

SingleDealerAssortment.mean_dols_per_archetype — Method

mean_dols_per_archetype(
    solution::Solution
) -> Vector{Float64}

Compute the mean days on lot per archetype of a solution.

SingleDealerAssortment.mean_gap_model — Method

mean_gap_model(Y_gap, model) -> Any

SingleDealerAssortment.mean_loss — Method

mean_loss(loss, model, X, Y) -> Any

SingleDealerAssortment.mean_replenishment_per_archetype — Method

mean_replenishment_per_archetype(
    solution::Solution
) -> Vector{Float64}

Compute the mean replenishment per archetype of a solution.

SingleDealerAssortment.mean_sales_per_archetype — Method

mean_sales_per_archetype(
    solution::Solution
) -> Vector{Float64}

Compute the mean sales per archetype of a solution.

SingleDealerAssortment.mean_stock_per_archetype — Method

mean_stock_per_archetype(
    solution::Solution
) -> Vector{Float64}

Compute the mean stock per archetype of a solution.

SingleDealerAssortment.merge_dataset — Method

merge_dataset(
    dataset_1::Dataset,
    dataset_2::Dataset
) -> Dataset

Helper function to merge two datasets.

SingleDealerAssortment.normalize_data! — Method

normalize_data!(X, μ, σ)

Standardize each feature of X by centering and reducing with μ and σ.

SingleDealerAssortment.order_policies! — Method

order_policies!(results::Results) -> Vector{String}

Order the policies in results according to POLICY_ORDER for ploting.

SingleDealerAssortment.print_costs — Method

print_costs(results::Results)

Print costs for each policy.

SingleDealerAssortment.print_gaps — Method

print_gaps(results::Results)

Print gaps to the anticipative bound for each policy.

SingleDealerAssortment.reduce_data! — Method

reduce_data!(X, σ)

Reduce X with σ, without centering it.

SingleDealerAssortment.replenishment_problem — Method

replenishment_problem(Θ; x, y_true, model_builder)

CO oracle with θ and η as cost vectors.

SingleDealerAssortment.sample_scenario — Method

sample_scenario(
    T,
    customer_choice_model,
    static_utilities,
    seed;
    temp
) -> Scenario

Sample a scenario given the customer choice model and static utilities.

SingleDealerAssortment.solve_anticipative — Method

solve_anticipative(
    instance::Instance,
    scenario::Scenario;
    time_limit,
    model_builder
) -> Union{Tuple{Int64, Nothing}, Tuple{Any, Solution}}

Solve the anticipative problem for a given instance and scenario.

SingleDealerAssortment.train_model — Method

train_model(
    initial_model,
    training_dataset::Dataset,
    test_dataset::Dataset;
    nb_epochs,
    ε,
    nb_samples,
    seed,
    lr,
    logger,
    log_dir,
    nb_max_epoch_without_improvement,
    early_stopping,
    with_eta,
    log_interval
) -> TrainingResults

Training loop for the COAML model.

SingleDealerAssortment.train_setup — Method

train_setup(
    initial_model,
    ε,
    nb_samples,
    seed,
    lr;
    with_eta
) -> NamedTuple{(:loss, :model, :opt), <:Tuple{InferOpt.FenchelYoungLoss{O} where O<:(InferOpt.PerturbedOracle{InferOpt.AdditivePerturbation{Distributions.Normal{Float64}}}), Any, Optimisers.Adam}}

SingleDealerAssortment.transform_data_for_training_set — Method

transform_data_for_training_set(
    solution::Solution
) -> Tuple{Vector{Matrix{Float32}}, Vector{Matrix{Float32}}}

Transform data into training set format. This function is used to create the oracle labels for y and z variables.